moved theme options to separate file
