+++ components/navigation/SlideMenu.tsx
@@ -1,25 +1,14 @@
 import * as React from "react";
-import { styled, useTheme } from "@mui/material/styles";
-import Box from "@mui/material/Box";
+import { styled, alpha } from "@mui/material/styles";
 import Drawer from "@mui/material/Drawer";
-import CssBaseline from "@mui/material/CssBaseline";
-import MuiAppBar, { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar";
-import Toolbar from "@mui/material/Toolbar";
 import List from "@mui/material/List";
-import Typography from "@mui/material/Typography";
 import Divider from "@mui/material/Divider";
 import IconButton from "@mui/material/IconButton";
-import MenuIcon from "@mui/icons-material/Menu";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import ChevronRightIcon from "@mui/icons-material/ChevronRight";
 import ListItem from "@mui/material/ListItem";
 import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
-import InboxIcon from "@mui/icons-material/MoveToInbox";
-import MailIcon from "@mui/icons-material/Mail";
-
-const drawerWidth = 240;
+import theme from "@/theme/theme";
 
 const DrawerHeader = styled("div")(({ theme }) => ({
     display: "flex",
@@ -28,6 +17,20 @@ const DrawerHeader = styled("div")(({ theme }) => ({
     // necessary for content to be below app bar
     ...theme.mixins.toolbar,
     justifyContent: "flex-end",
+}));
+
+const drawerWidth = 300;
+const StyledSlider = styled(Drawer)(({ theme }) => ({
+    width: drawerWidth,
+    background: "black",
+    border: "none",
+    "& .MuiDrawer-paper": {
+        width: drawerWidth,
+        boxSizing: "border-box",
+        height: "94%",
+        boxShadow: "",
+        border: "none",
+    },
 }));
 
 interface SlideMenuProps {
@@ -40,19 +43,7 @@ const SlideMenu: React.FC<SlideMenuProps> = ({
     handleslideMenuClose,
 }) => {
     return (
-        <Drawer
-            sx={{
-                width: drawerWidth,
-                flexShrink: 0,
-                "& .MuiDrawer-paper": {
-                    width: drawerWidth,
-                    boxSizing: "border-box",
-                },
-            }}
-            variant="persistent"
-            anchor="left"
-            open={slideMenuState}
-        >
+        <StyledSlider variant="persistent" anchor="left" open={slideMenuState}>
             <DrawerHeader>
                 <IconButton onClick={handleslideMenuClose}>
                     <ChevronLeftIcon />
@@ -63,15 +54,12 @@ const SlideMenu: React.FC<SlideMenuProps> = ({
                 {["Account", "Services", "Reviews"].map((text, index) => (
                     <ListItem key={text} disablePadding>
                         <ListItemButton>
-                            <ListItemIcon>
-                                {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-                            </ListItemIcon>
                             <ListItemText primary={text} />
                         </ListItemButton>
                     </ListItem>
                 ))}
             </List>
-        </Drawer>
+        </StyledSlider>
     );
 };
